feat: add search to page
