+++ packages/cli/types/main.d.ts
@@ -1,8 +1,8 @@
 import { DevServerConfig } from '@web/dev-server';
 import { MetaPlugin } from 'plugins-manager';
 // import { CheckHtmlLinksCliOptions } from 'check-html-links/dist-types/types/main';
-import { ImagePreset, RocketPreset } from './preset.js';
-export { ImagePreset, RocketPreset };
+import { ImagePreset, FullRocketPreset } from './preset.js';
+export { ImagePreset, FullRocketPreset as RocketPreset };
 
 import { Command } from 'commander';
 import { RocketCli } from '../src/RocketCli.js';
@@ -15,8 +15,8 @@ type PresetKeys =
   | 'setupCliPlugins'
   | 'setupEnginePlugins';
 
-export interface FullRocketCliOptions extends Pick<RocketPreset, PresetKeys> {
-  presets: Array<RocketPreset>;
+export interface FullRocketCliOptions extends Pick<FullRocketPreset, PresetKeys> {
+  presets: Array<FullRocketPreset>;
   // pathPrefix: string;
   serviceWorkerName: string;
   serviceWorkerSourcePath: string;
