+++ packages/launch/css/layout.css
@@ -200,6 +200,21 @@ body[layout^='layout-home'] #main-header a:hover {
   order: 3;
 }
 
+#main-header .search:not(:defined) {
+  height: 39px;
+  border: 1px solid var(--rocket-search-input-border-color, #dfe1e5);
+  box-shadow: none;
+  border-radius: var(--rocket-search-input-border-radius, 24px);
+  padding: 5px 0px;
+  width: 272px;
+  background: #fff;
+}
+
+#main-header .search:not(:defined) svg {
+  margin-left: 15px;
+  margin-top: 1px;
+}
+
 #main-header .logo-link span {
   font-size: 19px;
   font-weight: bold;
