+++ packages/search/custom-elements-manifest.config.js
@@ -1,13 +1,13 @@
-import { readmePlugin } from 'cem-plugin-readme';
-import { fileURLToPath } from 'url';
-import { dirname } from 'path';
+// import { readmePlugin } from 'cem-plugin-readme';
+// import { fileURLToPath } from 'url';
+// import { dirname } from 'path';
 
-export default {
-  globs: ['src/(RocketSearch|RocketSearchCombobox|RocketSearchOption).js'],
-  plugins: [
-    readmePlugin({
-      from: dirname(fileURLToPath(import.meta.url)),
-      header: 'README.head.md',
-    }),
-  ],
-};
+// export default {
+//   globs: ['src/(RocketSearch|RocketSearchCombobox|RocketSearchOption).js'],
+//   plugins: [
+//     readmePlugin({
+//       from: dirname(fileURLToPath(import.meta.url)),
+//       header: 'README.head.md',
+//     }),
+//   ],
+// };
