--- packages/search/node.js
@@ -1,4 +1,0 @@
-export { getIdBlocksOfHtml } from './src/getIdBlocksOfHtml.js';
-export { RocketSearchPlugin } from './src/RocketSearchPlugin.js';
-
-export { rocketSearch } from './preset/rocketSearch.js';
