--- packages/search/preset/_assets/_static/search/noscript.css
@@ -1,3 +1,0 @@
-rocket-search {
-  display: none;
-}
