--- packages/search/preset/_assets/scripts/define-rocket-search.js
@@ -1,1 +1,0 @@
-import '@rocket/search/rocket-search.js';
