--- packages/search/preset/_includes/_joiningBlocks/head/43-search-noscript.njk
@@ -1,1 +1,0 @@
-<noscript><link rel="stylesheet" href="{{ '/_assets/_static/search/noscript.css' | asset | url }}"></noscript>
