+++ packages/search/src/define.js
@@ -1,3 +1,3 @@
-import { RocketSearch } from './src/RocketSearch.js';
+import { RocketSearch } from './RocketSearch.js';
 
 customElements.define('rocket-search', RocketSearch);
