+++ packages/search/src/index.components.js
@@ -1,0 +1,3 @@
+export const searchComponents = {
+  // 'rocket-search': () => import('./RocketSearch.js').then(m => m.RocketSearch),
+};
