+++ packages/search/src/index.node.js
@@ -1,0 +1,4 @@
+// export { getIdBlocksOfHtml } from './getIdBlocksOfHtml.js';
+// export { RocketSearchPlugin } from './RocketSearchPlugin.js';
+
+export { presetRocketSearch } from './presetRocketSearch.js';
