+++ packages/search/src/index.web.js
@@ -1,0 +1,1 @@
+export { RocketSearch } from './RocketSearch.js';
