+++ packages/search/test-node/fixtures/01-single-page/config/rocket.config.js
@@ -1,0 +1,5 @@
+import { presetRocketSearch } from '@rocket/search';
+
+export default /** @type {import('@rocket/cli').RocketCliOptions} */ ({
+  presets: [presetRocketSearch()],
+});
