+++ packages/search/test-node/fixtures/01-single-page/site/pages/index.rocket.js
@@ -1,0 +1,13 @@
+/* START - Rocket auto generated - do not touch */
+export const sourceRelativeFilePath = 'index.rocket.js';
+/* END - Rocket auto generated - do not touch */
+
+import { html } from 'lit';
+
+export default () => html`
+  <h1>Home</h1>
+  <p>My home is my castle</p>
+
+  <h2 id="about">About</h2>
+  <p>I am human</p>
+`;
