+++ packages/search/test-node/fixtures/01-single-page/site/pages/pageTreeData.rocketGenerated.json
@@ -1,0 +1,16 @@
+{
+  "h1": "Home",
+  "headlinesWithId": [
+    {
+      "text": "About",
+      "id": "about",
+      "level": 2
+    }
+  ],
+  "name": "Home",
+  "menuLinkText": "Home",
+  "url": "/",
+  "outputRelativeFilePath": "index.html",
+  "sourceRelativeFilePath": "index.rocket.js",
+  "level": 0
+}
