--- packages/search/web.js
@@ -1,1 +1,0 @@
-export { RocketSearch } from './src/RocketSearch.js';
