+++ site/pages/index.rocket.js
@@ -7,6 +7,7 @@ export { html, setupUnifiedPlugins, components };
 import { footerMenu } from './__shared/footerMenu.js';
 import { pageTree } from './__shared/pageTree.js';
 import { LayoutHome } from '@rocket/launch';
+import { search } from './__shared/search.js';
 
 export const description =
   'Rocket enables everyone to code a website. Use an existing theme or create your own. Be fast by server rendering web components with little to no JavaScript.';
@@ -15,6 +16,7 @@ export const subTitle = 'Everyone can code a website';
 export const layout = new LayoutHome({
   pageTree,
   footerMenu,
+  header__40: search,
   titleWrapperFn: () => 'Welcome to Rocket',
   slogan: 'Everyone can code a website.',
   callToActionItems: [
