+++ workspace-packages.mjs
@@ -3,6 +3,7 @@ const packages = [
   { name: 'engine', type: 'js', environment: 'node-esm' },
   { name: 'plugins-manager', type: 'js', environment: 'node-esm' },
   { name: 'launch', type: 'js', environment: 'node-esm' },
+  { name: 'search', type: 'js', environment: 'node-esm' },
   // { name: 'check-html-links', type: 'js', environment: 'node-esm' },
   // { name: 'drawer', type: 'js', environment: 'browser' },
   { name: 'mdjs-core', type: 'js', environment: 'node' },
